+++ src/components/tx-item.tsx
